requirements.txt
torch
torchvision
tensorboard
tqdm
matplotlib
PyYAML